fix: move render_ai_payload tests to cfg(test) module, add structured=None proposal clearing test (#54)
- Restore render_ai_payload as private fn
- Move 5 formatter tests into #[cfg(test)] mod in src/application/mod.rs
- Add Application-level test: structured=None clears skill proposals
